+++ packages/essential_dart/example/essential_dart_example.dart
@@ -1,6 +1,1 @@
-import 'package:essential_dart/essential_dart.dart';
-
-void main() {
-  final awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}
+void main() {}