+++ packages/essential_dart/lib/src/types.dart
@@ -1,0 +1,7 @@
+import 'dart:async';
+
+typedef SyncComputation<T> = T Function();
+
+typedef AsyncComputation<T> = Future<T> Function();
+
+typedef Computation<T> = FutureOr<T> Function();